requirements.txt
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2==2.9.11
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
stripe==5.0.0
psycopg2-binary==2.9.11
langchain==1.2.0
openai==2.17.0
langchain-community==0.4.1
langchain-openai==1.1.7
chromadb==1.5.0
sentence-transformers==5.2.2
torch==2.9.0
transformers==5.1.0
pandas==3.0.0
numpy==2.4.2
openpyxl==3.1.5
# redis==6.2.13
pydantic[email]==2.12.5